+ renaming queue and class to match rest of microservices
