+++ Calculator/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 server.port=8080
-#spring.rabbitmq.listener.simple.auto-startup=false
+#spring.rabbitmq.receiver.simple.auto-startup=false
 spring.application.name=calculator
 #spring.config.import=configserver:http://localhost:8888
 spring.config.import=optional:configserver:http://configserver:8888
